lotta changes
* renamed user-service to api-service
* made it a resource server using keycloak
* removed unnecessary files/packages from api-service, also removed dependencies.
