+++ api-service/src/main/java/com/rafantasy/userservice/UserServiceApplication.java
@@ -1,0 +1,18 @@
+package com.rafantasy.userservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+
+@ConfigurationPropertiesScan
+@SpringBootApplication(exclude = { SecurityAutoConfiguration.class })
+@EnableWebSecurity
+public class UserServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(UserServiceApplication.class, args);
+	}
+
+}
