+++ api-service/src/main/java/com/rafantasy/userservice/common/Constants.java
@@ -1,0 +1,11 @@
+package com.rafantasy.userservice.common;
+
+/**
+ * This class contains all of the constants.
+ */
+public class Constants {
+    public static final String POC_APP_NO_REPLY = "noreply@pocApp.com";
+    public static final String ROLE_NOT_FOUND = "role.not.found";
+    public static final String USER_NOT_FOUND = "user.not.found";
+    public static final String WRONG_PASSWORD_PROVIDED = "wrong.password.provided";
+}
