+++ api-service/src/main/java/com/rafantasy/userservice/user/controller/UserController.java
@@ -1,0 +1,21 @@
+package com.rafantasy.userservice.user.controller;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.*;
+
+/**
+ * User Controller.
+ */
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/api/users")
+@Slf4j
+public class UserController {
+
+    @GetMapping
+    public String getAllUsers() {
+        return "works";
+    }
+
+}
