+++ api-service/src/main/resources/application.yaml
@@ -1,0 +1,46 @@
+spring:
+  application:
+    name: api-service
+
+  security:
+    oauth2:
+      resourceserver:
+        jwt:
+          issuer-uri: http://localhost:8180/auth/realms/Rafantasy
+
+  datasource:
+    url: jdbc:postgresql://localhost:5430/rafantasy?applicationName=rafantasy-user-service
+    username: postgres
+    password: password
+    driver-class-name: org.postgresql.Driver
+    platform: postgres
+    initialization-mode: never
+    tomcat:
+      max-wait: 60000
+
+  jpa:
+    hibernate:
+      ddl-auto: none
+    properties:
+      hibernate.implicit_naming_strategy: org.springframework.boot.orm.jpa.hibernate.SpringImplicitNamingStrategy
+      hibernate.physical_naming_strategy: org.springframework.boot.orm.jpa.hibernate.SpringPhysicalNamingStrategy
+
+  mail:
+    protocol: smtp
+    host: smtp.mailtrap.io
+    port: 2525
+    username: ${MAIL_TRAP_USERNAME_CRED}
+    password: ${MAIL_TRAP_PASSWORD_CRED}
+    properties:
+      mail:
+        smtp:
+          auth: true
+          connectiontimeout: 5000
+          timeout: 5000
+          writetimeout: 5000
+          starttls:
+            enable: true
+server:
+  error:
+    include-message: always
+  port: 8081
