--- user-service/src/main/java/com/rafantasy/userservice/UserServiceApplication.java
@@ -1,16 +1,0 @@
-package com.rafantasy.userservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
-import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
-
-@ConfigurationPropertiesScan
-@SpringBootApplication(exclude = { SecurityAutoConfiguration.class })
-public class UserServiceApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(UserServiceApplication.class, args);
-	}
-
-}
