--- user-service/src/main/java/com/rafantasy/userservice/domain/dto/ChangePasswordDTO.java
@@ -1,23 +1,0 @@
-package com.rafantasy.userservice.domain.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.experimental.Accessors;
-
-import javax.persistence.Access;
-import javax.validation.constraints.NotBlank;
-
-@Data
-@Accessors(chain = true)
-@AllArgsConstructor
-@NoArgsConstructor
-public class ChangePasswordDTO {
-
-    @NotBlank(message = "old.password.empty")
-    private String oldPassword;
-
-    @NotBlank(message = "new.password.empty")
-    private String newPassword;
-
-}
