--- user-service/src/main/java/com/rafantasy/userservice/domain/model/Role.java
@@ -1,29 +1,0 @@
-package com.rafantasy.userservice.domain.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.experimental.Accessors;
-
-import javax.persistence.*;
-
-/**
- * Role Model.
- */
-@Entity
-@Table(name = "role")
-@Getter
-@Setter
-@Accessors(chain = true)
-@AllArgsConstructor
-@NoArgsConstructor
-public class Role {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
-    @Enumerated(EnumType.STRING)
-    private RoleType roleType;
-}
