--- user-service/src/main/java/com/rafantasy/userservice/domain/model/RoleType.java
@@ -1,8 +1,0 @@
-package com.rafantasy.userservice.domain.model;
-
-/**
- * Role Types.
- */
-public enum RoleType {
-    ROLE_ADMIN, ROLE_USER
-}
