--- user-service/src/main/java/com/rafantasy/userservice/domain/service/RoleRepository.java
@@ -1,22 +1,0 @@
-package com.rafantasy.userservice.domain.service;
-
-import com.rafantasy.userservice.domain.model.Role;
-import com.rafantasy.userservice.domain.model.RoleType;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-/**
- * Role repository.
- */
-@Repository
-public interface RoleRepository extends JpaRepository<Role, Integer> {
-
-    /**
-     * To find a Role entry when provided with its role type.
-     * @param roleType The role type.
-     * @return Role entry, if present.
-     */
-    Optional<Role> findByRoleType(RoleType roleType);
-}
