--- user-service/src/main/java/com/rafantasy/userservice/security/KeyCloakSecurityConfig.java
@@ -1,54 +1,0 @@
-//package com.rafantasy.userservice.security;
-//
-//import org.keycloak.adapters.springboot.KeycloakSpringBootConfigResolver;
-//import org.keycloak.adapters.springsecurity.KeycloakSecurityComponents;
-//import org.keycloak.adapters.springsecurity.authentication.KeycloakAuthenticationProvider;
-//import org.keycloak.adapters.springsecurity.config.KeycloakWebSecurityConfigurerAdapter;
-//import org.springframework.beans.factory.annotation.Autowired;
-//import org.springframework.context.annotation.Bean;
-//import org.springframework.context.annotation.ComponentScan;
-//import org.springframework.context.annotation.Configuration;
-//import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-//import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-//import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-//import org.springframework.security.core.authority.mapping.SimpleAuthorityMapper;
-//import org.springframework.security.core.session.SessionRegistryImpl;
-//import org.springframework.security.web.authentication.session.RegisterSessionAuthenticationStrategy;
-//import org.springframework.security.web.authentication.session.SessionAuthenticationStrategy;
-//
-//@Configuration
-//@EnableWebSecurity
-//@ComponentScan(basePackageClasses = KeycloakSecurityComponents.class)
-//public class KeyCloakSecurityConfig extends KeycloakWebSecurityConfigurerAdapter {
-//
-//    @Autowired
-//    public void configureGlobal(AuthenticationManagerBuilder auth) throws Exception {
-//
-//        KeycloakAuthenticationProvider keycloakAuthenticationProvider
-//                = keycloakAuthenticationProvider();
-//        keycloakAuthenticationProvider.setGrantedAuthoritiesMapper(
-//                new SimpleAuthorityMapper());
-//        auth.authenticationProvider(keycloakAuthenticationProvider);
-//    }
-//
-//    @Bean
-//    public KeycloakSpringBootConfigResolver KeycloakConfigResolver() {
-//        return new KeycloakSpringBootConfigResolver();
-//    }
-//
-//    @Bean
-//    @Override
-//    protected SessionAuthenticationStrategy sessionAuthenticationStrategy() {
-//        return new RegisterSessionAuthenticationStrategy(
-//                new SessionRegistryImpl());
-//    }
-//
-//    @Override
-//    protected void configure(HttpSecurity http) throws Exception {
-//        super.configure(http);
-//        http.cors().and().csrf().disable().authorizeRequests()
-//                .antMatchers("/api/users/**").hasRole("user")
-//                .anyRequest()
-//                .permitAll();
-//    }
-//}
