--- user-service/src/main/java/com/rafantasy/userservice/security/UserDetailsImpl.java
@@ -1,56 +1,0 @@
-//package com.rafantasy.userservice.security;
-//
-//import com.rafantasy.userservice.domain.model.User;
-//import lombok.RequiredArgsConstructor;
-//import org.springframework.security.core.GrantedAuthority;
-//import org.springframework.security.core.authority.SimpleGrantedAuthority;
-//import org.springframework.security.core.userdetails.UserDetails;
-//
-//import java.util.ArrayList;
-//import java.util.Collection;
-//
-//@RequiredArgsConstructor
-//public class UserDetailsImpl implements UserDetails {
-//
-//    private final User user;
-//
-//    @Override
-//    public Collection<? extends GrantedAuthority> getAuthorities() {
-//
-//        var authorities = new ArrayList<GrantedAuthority>();
-//        user
-//                .getRoles()
-//                .forEach(grantedAuthority -> authorities.add(new SimpleGrantedAuthority(grantedAuthority.getRoleType().toString())));
-//        return authorities;
-//    }
-//
-//    @Override
-//    public String getPassword() {
-//        return user.getPassword();
-//    }
-//
-//    @Override
-//    public String getUsername() {
-//        return user.getUsername();
-//    }
-//
-//    @Override
-//    public boolean isAccountNonExpired() {
-//        return true;
-//    }
-//
-//    @Override
-//    public boolean isAccountNonLocked() {
-//        return user.getIsNonLocked();
-//    }
-//
-//    @Override
-//    public boolean isCredentialsNonExpired() {
-//        return true;
-//    }
-//
-//    @Override
-//    public boolean isEnabled() {
-//        return user.getIsActive();
-//    }
-//}
