--- user-service/src/main/java/com/rafantasy/userservice/security/dto/ForgotPasswordDTO.java
@@ -1,17 +1,0 @@
-package com.rafantasy.userservice.security.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.experimental.Accessors;
-
-@Data
-@Accessors(chain = true)
-@AllArgsConstructor
-@NoArgsConstructor
-public class ForgotPasswordDTO {
-
-    private String username;
-
-    private String email;
-}
