--- user-service/src/main/java/com/rafantasy/userservice/security/dto/LoginResponseDTO.java
@@ -1,35 +1,0 @@
-package com.rafantasy.userservice.security.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.experimental.Accessors;
-
-@Getter
-@Setter
-@Accessors(chain = true)
-@AllArgsConstructor
-@NoArgsConstructor
-public class LoginResponseDTO {
-
-    @JsonProperty("token_type")
-    private String tokenType;
-
-    @JsonProperty("access_token")
-    private String accessToken;
-
-    @JsonProperty("expires_in")
-    private Long expiresIn;
-
-    @JsonProperty("refresh_token")
-    private String refreshToken;
-
-    @JsonProperty("refresh_expires_in")
-    private Long refreshExpiresIn;
-
-    @JsonProperty("session_state")
-    private String sessionState;
-
-}
