--- user-service/src/main/java/com/rafantasy/userservice/security/dto/RefreshTokenRequestDTO.java
@@ -1,12 +1,0 @@
-package com.rafantasy.userservice.security.dto;
-
-import lombok.Data;
-
-import javax.validation.constraints.NotBlank;
-
-@Data
-public class RefreshTokenRequestDTO {
-
-    @NotBlank
-    private String refreshToken;
-}
