--- user-service/src/main/java/com/rafantasy/userservice/security/enums/TokenType.java
@@ -1,5 +1,0 @@
-package com.rafantasy.userservice.security.enums;
-
-public enum TokenType {
-    signup, forgot_password
-}
