--- user-service/src/main/java/com/rafantasy/userservice/security/model/RefreshToken.java
@@ -1,34 +1,0 @@
-package com.rafantasy.userservice.security.model;
-
-import com.rafantasy.userservice.domain.model.User;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import lombok.experimental.Accessors;
-
-import javax.persistence.*;
-import java.time.Instant;
-
-@Entity
-@Table(name = "refresh_token")
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Accessors(chain = true)
-public class RefreshToken {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @OneToOne
-    @JoinColumn(name = "user_id", referencedColumnName = "id")
-    private User user;
-
-    @Column(nullable = false, unique = true)
-    private String token;
-
-    @Column(nullable = false)
-    private Instant expiresIn;
-
-}
