--- user-service/src/main/java/com/rafantasy/userservice/security/service/AuthService.java
@@ -1,9 +1,0 @@
-package com.rafantasy.userservice.security.service;
-
-import com.rafantasy.userservice.security.dto.LoginRequestDTO;
-import com.rafantasy.userservice.security.dto.LoginResponseDTO;
-
-public interface AuthService {
-
-    LoginResponseDTO login(LoginRequestDTO loginRequestDTO);
-}
