--- user-service/src/main/java/com/rafantasy/userservice/security/service/RefreshTokenRepository.java
@@ -1,16 +1,0 @@
-package com.rafantasy.userservice.security.service;
-
-import com.rafantasy.userservice.domain.model.User;
-import com.rafantasy.userservice.security.model.RefreshToken;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
-
-    Optional<RefreshToken> findByToken(String token);
-
-    void deleteByUser(User user);
-}
