--- user-service/src/main/resources/application.yaml
@@ -1,61 +1,0 @@
-spring:
-  application:
-    name: user-service
-  security:
-    oauth2:
-      resourceserver:
-        opaque:
-          introspection-uri: http://localhost:8180/auth/realms/Rafantasy/protocol/openid-connect/token/introspect
-          introspection-client-id: user-service
-          introspection-client-secret: 0eb75b3c-d42c-483b-acf9-cee71c718926
-
-  datasource:
-    url: jdbc:postgresql://localhost:5430/rafantasy?applicationName=rafantasy-user-service
-    username: postgres
-    password: password
-    driver-class-name: org.postgresql.Driver
-    platform: postgres
-    initialization-mode: never
-    tomcat:
-      max-wait: 60000
-
-  jpa:
-    hibernate:
-      ddl-auto: none
-    properties:
-      hibernate.implicit_naming_strategy: org.springframework.boot.orm.jpa.hibernate.SpringImplicitNamingStrategy
-      hibernate.physical_naming_strategy: org.springframework.boot.orm.jpa.hibernate.SpringPhysicalNamingStrategy
-
-  mail:
-    protocol: smtp
-    host: smtp.mailtrap.io
-    port: 2525
-    username: ${MAIL_TRAP_USERNAME_CRED}
-    password: ${MAIL_TRAP_PASSWORD_CRED}
-    properties:
-      mail:
-        smtp:
-          auth: true
-          connectiontimeout: 5000
-          timeout: 5000
-          writetimeout: 5000
-          starttls:
-            enable: true
-server:
-  error:
-    include-message: always
-  port: 8081
-
-
-user-service-properties:
-  jwt-secret: ${JWT_SECRET}
-  jwt-expiry-time-in-ms: ${JWT_EXPIRY_TIME_IN_MS}
-  front-end-url: ${FRONT_END_URL}
-  refreshTokenExpiryTimeInMs: 3600000
-
-
-#keycloak:
-#  auth-server-url: http://localhost:8180/auth
-#  realm: Rafantasy
-#  resource: user-service
-#  public-client: true
